pyproject.toml
[project]
name = "openapi-to-mcp"
version = "0.1.0"
description = "A tool to convert OpenAPI specifications to Layer MCP format"
readme = "README.md"
authors = [
    { name = "andrewlayer", email = "andrewh@buildwithlayer.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.1",
    "pydantic>=2.6.1",
    "typer>=0.9.0",
    "mcp[cli]>=1.3.0",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

[project.scripts]
openapi-to-mcp = "openapi_to_mcp.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.urls]
Repository = "https://github.com/buildwithlayer/openapi-to-mcp"

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=openapi_to_mcp --cov-report=term-missing"
